+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ## [Unreleased]
 
+### ⚡ Changes
+
+- Don't give a false impression that the Glyph Button interaction works when there's no associated React Context available.
+
 ## [0.3.0] - 2026-03-03
 
 ### ❗ Breaking Changes