+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ## [Unreleased]
 
+## [0.4.0] - 2026-04-28
+
 ### ⚡ Changes
 
 - Don't give a false impression that the Glyph Button interaction works when there's no associated React Context available.
@@ -97,7 +99,8 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 Add section to make `release-it` not complain that this is missing. 2025-09-08 is when we really started working on this repository.
 
-[Unreleased]: https://github.com/MissingCore/music-glyph-toys/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/MissingCore/music-glyph-toys/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/MissingCore/music-glyph-toys/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/MissingCore/music-glyph-toys/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/MissingCore/music-glyph-toys/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/MissingCore/music-glyph-toys/compare/v0.2.0...v0.2.1